pyproject.toml
[tool.poetry]
name = "logviewer2"
version = "0.1.0"
description = ""
authors = ["Nadie <hackerjef@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^2.2.2"
gevent = "^22.10.2"
natural = "^0.2.0"
dateutils = "^0.6.12"
click = "^8.1.3"
python-dotenv = "^0.21.0"
requests-oauthlib = "^1.3.1"
pymongo = {extras = ["srv"], version = "^4.3.3"}
sentry-sdk = {extras = ["flask"], version = "^1.11.0"}
flask-discord = {git = "https://github.com/weibeu/Flask-Discord.git", rev = "v9"}
gunicorn = "^20.1.0"
psycogreen = "^1.0.2"

[tool.poetry.scripts]
web = 'scripts:web'
webd = 'scripts:webd'
secret = 'scripts:secret'
gensecret = 'scripts:gensecret'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
